sharing/requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
Pillow==8.4.0
